requirements.txt
streamlit
faster-whisper
pyannote.audio==3.1.1
transformers
librosa
soundfile
torch
python-dotenv
pydantic
pytest
numpy
vaderSentiment
